+++ public/app/api/clients/folder/baseAPI.ts
@@ -5,7 +5,7 @@ import { getAPIBaseURL } from 'app/api/utils';
 
 export const BASE_URL = getAPIBaseURL('folder.grafana.app', 'v0alpha1');
 
-export const baseAPI = createApi({
+export const api = createApi({
   reducerPath: 'folderAPI',
   baseQuery: createBaseQuery({
     baseURL: BASE_URL,
