+++ public/app/api/clients/iam/baseAPI.ts
@@ -5,7 +5,7 @@ import { getAPIBaseURL } from '../../utils';
 
 export const BASE_URL = getAPIBaseURL('iam.grafana.app', 'v0alpha1');
 
-export const iamAPI = createApi({
+export const api = createApi({
   baseQuery: createBaseQuery({ baseURL: BASE_URL }),
   reducerPath: 'iamAPI',
   endpoints: () => ({}),
