+++ public/app/api/clients/iam/index.ts
@@ -1,3 +1,5 @@
 import { generatedAPI } from './endpoints.gen';
 
+export const iamAPI = generatedAPI.enhanceEndpoints({});
+
 export const { useGetDisplayMappingQuery } = generatedAPI;
